update date parsing
