+++ tests/data/dummy_connection/connection.yaml
@@ -30,13 +30,13 @@ is_abstract: false
 cert_requests:
 - identifier: cert_id_1
   ledger_id: some_ledger_id
-  not_after: '2020-01-02T00:00:00'
-  not_before: '2020-01-01T00:00:00'
+  not_after: '2020-01-02'
+  not_before: '2020-01-01'
   path: some/path_1
   public_key: key_id
 - identifier: cert_id_2
   ledger_id: some_ledger_id
-  not_after: '2020-01-02T00:00:00'
-  not_before: '2020-01-01T00:00:00'
+  not_after: '2020-01-02'
+  not_before: '2020-01-01'
   path: some/path_2
   public_key: '0xABCDEF123456'
