+++ tests/data/hashes.csv
@@ -1,6 +1,6 @@
 dummy_author/agents/dummy_aea,QmPWxJVM9xnkRFtB41Wr1Xi2NGKpeRgDXUzbAzA8rferGh
 dummy_author/skills/dummy_skill,QmSmzcjGNhiz2KpSE6wSi19AeesCVjzupJm1NpXYJVnJuZ
-fetchai/connections/dummy_connection,QmTCvVMLKiJFWHF3xb6JdGmz3HBdkiYuNrADd2WGnaDeV5
+fetchai/connections/dummy_connection,QmWCzxyUTivdaDiouSepKwjyomyshdzocSHDStXG16ZHkL
 fetchai/contracts/dummy_contract,QmXhMKRF5upco1efaC9oNcesStEMkPWKhBUv6CK9Vngk2S
 fetchai/protocols/t_protocol,QmNhzDCznorQpySrvKt4F4sQFXjzF1ZecEd5inixiNavDH
 fetchai/protocols/t_protocol_no_ct,QmWwMwHxvuPZJaMh44Bd1d5aBzYiogep8TJswTKkRwLfBu
